renaming the modules is not a good idea
