+++ ptypy/accelerate/cuda_pycuda/engines/projectional_pycuda_stream.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
 """
-Difference Map reconstruction engine.
+Difference Map reconstruction engine for NVIDIA GPUs.
+
+This engine uses three streams, one for the compute queue and one for each I/O queue.
+Events are used to synchronize download / compute/ upload. we cannot manipulate memory
+for each loop over the state vector, a certain number of memory sections is preallocated
+and reused.
 
 This file is part of the PTYPY package.
 
@@ -12,111 +17,66 @@ import numpy as np
 import time
 from pycuda import gpuarray
 import pycuda.driver as cuda
+from pycuda.tools import DeviceMemoryPool
 
 from ptypy import utils as u
-from ptypy.utils.verbose import logger, log
+from ptypy.utils.verbose import log, logger
 from ptypy.utils import parallel
 from ptypy.engines import register
 from ptypy.engines.projectional import DMMixin, RAARMixin
 from . import projectional_pycuda
-from ..mem_utils import GpuDataManager
 
-# factor how many more exit waves we wanna keep on GPU compared to 
-# ma / mag data
+from ..mem_utils import make_pagelocked_paired_arrays as mppa
+from ..mem_utils import GpuDataManager2
+
 EX_MA_BLOCKS_RATIO = 2
-MAX_STREAMS = 500   # max number of streams to use
 MAX_BLOCKS = 99999  # can be used to limit the number of blocks, simulating that they don't fit
+#MAX_BLOCKS = 3  # can be used to limit the number of blocks, simulating that they don't fit
 
-__all__ = ['DM_pycuda_streams']
-
-class GpuStreamData:
-    def __init__(self, ex_data, ma_data, mag_data):
-        self.queue = cuda.Stream()
-        self.ex_data = ex_data
-        self.ma_data = ma_data
-        self.mag_data = mag_data
-        self.ev_compute = None
-
-    def end_compute(self):
-        """
-        called at end of kernels using shared data (aux or probe),
-        to mark when computing is done and it can be re-used
-        """
-        self.ev_compute = cuda.Event()
-        self.ev_compute.record(self.queue)
-        return self.ev_compute
-
-    def start_compute(self, prev_event):
-        """
-        called at start of kernels using shared data (aux or probe),
-        to wait for previous use of this data is finished
-        """
-
-        if prev_event is not None:
-            self.queue.wait_for_event(prev_event)
-
-    def ex_to_gpu(self, dID, ex):
-        """
-        copy exit wave to GPU, but check first if it's already there
-        If not, but a previous block is on GPU, sync it back to the host
-        before overwriting it
-        """
-
-        return self.ex_data.to_gpu(ex, dID, self.queue)
-
-    def ma_to_gpu(self, dID, ma, mag):
-        """
-        Copy MA array to GPU
-        """
-        # wait for previous work on memory to complete
-        ma_gpu = self.ma_data.to_gpu(ma, dID, self.queue)
-        mag_gpu = self.mag_data.to_gpu(mag, dID, self.queue)
-        return ma_gpu, mag_gpu
-
-    def record_done_ex(self, dID):
-        """
-        Record when we're done with this stream, so that it can be re-used
-        """
-        self.ex_data.record_done(dID, self.queue)
-
-    def record_done_ma(self, dID):
-        self.ma_data.record_done(dID, self.queue)
-        self.mag_data.record_done(dID, self.queue)
-
-    def synchronize(self):
-        """
-        Wait for stream to finish its work
-        """
-        self.queue.synchronize()
+__all__ = ['DM_pycuda', 'RAAR_pycuda']
 
 
-class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pycuda):
+class _ProjectionEngine_pycuda(projectional_pycuda._ProjectionEngine_pycuda_nostream):
 
-    """
-    Defaults:
-
-    [fft_lib]
-    default = cuda
-    type = str
-    help = Choose the pycuda-compatible FFT module.
-    doc = One of:
-      - ``'reikna'`` : the reikna packaga (fast load, competitive compute for streaming)
-      - ``'cuda'`` : ptypy's cuda wrapper (delayed load, but fastest compute if all data is on GPU)
-      - ``'skcuda'`` : scikit-cuda (fast load, slowest compute due to additional store/load stages)
-    choices = 'reikna','cuda','skcuda'
-    userlevel = 2
-
-    """
-
-    def __init__(self, ptycho_parent, pars = None):
+    def __init__(self, ptycho_parent, pars=None):
 
         super().__init__(ptycho_parent, pars)
-        self.streams = None 
         self.ma_data = None
         self.mag_data = None
         self.ex_data = None
-        self.cur_stream = 0
-        self.stream_direction = 1
+
+    def engine_initialize(self):
+        super().engine_initialize()
+        self.qu_htod = cuda.Stream()
+        self.qu_dtoh = cuda.Stream()
+
+    def _setup_kernels(self):
+
+        super()._setup_kernels()
+        ex_mem = 0
+        mag_mem = 0
+        for scan, kern in self.kernels.items():
+            ex_mem = max(kern.aux.nbytes, ex_mem)
+            mag_mem = max(kern.FUK.gpu.fdev.nbytes, mag_mem)
+        ma_mem = mag_mem
+        mem = cuda.mem_get_info()[0]
+        blk = ex_mem * EX_MA_BLOCKS_RATIO + ma_mem + mag_mem
+        fit = int(mem - 200 * 1024 * 1024) // blk  # leave 200MB room for safety
+        if not fit:
+            log(1,"Cannot fit memory into device, if possible reduce frames per block. Exiting...")
+            self.context.pop()
+            self.context.detach()
+            raise SystemExit("ptypy has been exited.")
+
+        # TODO grow blocks dynamically
+        nex = min(fit * EX_MA_BLOCKS_RATIO, MAX_BLOCKS)
+        nma = min(fit, MAX_BLOCKS)
+        log(4, 'Free memory on device: %.2f GB' % (float(mem)/1e9))
+        log(4, 'PyCUDA max blocks fitting on GPU: exit arrays={}, ma_arrays={}'.format(nex, nma))
+        # reset memory or create new
+        self.ex_data = GpuDataManager2(ex_mem, 0, nex, True)
+        self.ma_data = GpuDataManager2(ma_mem, 0, nma, False)
+        self.mag_data = GpuDataManager2(mag_mem, 0, nma, False)
 
     def engine_prepare(self):
 
@@ -124,40 +84,16 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
 
         for name, s in self.ob.S.items():
             s.gpu = gpuarray.to_gpu(s.data)
-        # we use default mem_flags for ob/obn/pr/prn page-locking, as we are 
-        # operating on them on CPU as well after each iteration.
-        # Write-Combined memory (flags=4) is for write-only on CPU side,
-        # reads are really slow.
         for name, s in self.ob_buf.S.items():
-            # obb
-            d = s.data
-            s.data = cuda.pagelocked_empty(d.shape, d.dtype, order="C", mem_flags=0)
-            s.data[:] = d
-            s.gpu = gpuarray.to_gpu(s.data)
+            s.gpu, s.data = mppa(s.data)
         for name, s in self.ob_nrm.S.items():
-            # obn
-            d = s.data
-            s.data = cuda.pagelocked_empty(d.shape, d.dtype, order="c", mem_flags=0)
-            s.data[:] = d
-            s.gpu = gpuarray.to_gpu(s.data)
+            s.gpu, s.data = mppa(s.data)
         for name, s in self.pr.S.items():
-            # pr
-            d = s.data
-            s.data = cuda.pagelocked_empty(d.shape, d.dtype, order="C", mem_flags=0)
-            s.data[:] = d
-            s.gpu = gpuarray.to_gpu(s.data)
+            s.gpu, s.data = mppa(s.data)
         for name, s in self.pr_buf.S.items():
-            # pr
-            d = s.data
-            s.data = cuda.pagelocked_empty(d.shape, d.dtype, order="C", mem_flags=0)
-            s.data[:] = d
-            s.gpu = gpuarray.to_gpu(s.data)
+            s.gpu, s.data = mppa(s.data)
         for name, s in self.pr_nrm.S.items():
-            # prn
-            d = s.data
-            s.data = cuda.pagelocked_empty(d.shape, d.dtype, order="C", mem_flags=0)
-            s.data[:] = d
-            s.gpu = gpuarray.to_gpu(s.data)
+            s.gpu, s.data = mppa(s.data)
 
         use_tiles = (not self.p.probe_update_cuda_atomics) or (not self.p.object_update_cuda_atomics)
 
@@ -166,19 +102,20 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
             for name, s in self.ob_buf.S.items():
                 s.tmp = gpuarray.empty(s.gpu.shape, s.gpu.dtype)
 
-        ex_mem = ma_mem = mag_mem = 0
-        idlist = list(self.di.S.keys())
-        blocks = len(idlist)
-        for dID in idlist:
-            prep = self.diff_info[dID]
-            pID, oID, eID = prep.poe_IDs
-
+        # TODO : like the serialization this one is needed due to object reformatting
+        for label, d in self.di.storages.items():
+            prep = self.diff_info[d.ID]
             prep.addr_gpu = gpuarray.to_gpu(prep.addr)
             if use_tiles:
                 prep.addr2 = np.ascontiguousarray(np.transpose(prep.addr, (2, 3, 0, 1)))
                 prep.addr2_gpu = gpuarray.to_gpu(prep.addr2)
             if self.do_position_refinement:
                 prep.mangled_addr_gpu = prep.addr_gpu.copy()
+
+        for label, d in self.ptycho.new_data:
+            dID = d.ID
+            prep = self.diff_info[dID]
+            pID, oID, eID = prep.poe_IDs
 
             prep.ma_sum_gpu = gpuarray.to_gpu(prep.ma_sum)
             # prepare page-locked mems:
@@ -189,62 +126,29 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                 prep.error_state_gpu = gpuarray.empty_like(prep.err_fourier_gpu)
             ma = self.ma.S[dID].data.astype(np.float32)
             prep.ma = cuda.pagelocked_empty(ma.shape, ma.dtype, order="C", mem_flags=4)
-            prep.ma[:] = ma   
+            prep.ma[:] = ma
             ex = self.ex.S[eID].data
             prep.ex = cuda.pagelocked_empty(ex.shape, ex.dtype, order="C", mem_flags=4)
             prep.ex[:] = ex
             mag = prep.mag
             prep.mag = cuda.pagelocked_empty(mag.shape, mag.dtype, order="C", mem_flags=4)
             prep.mag[:] = mag
-            ex_mem = max(ex_mem, prep.ex.nbytes)
-            ma_mem = max(ma_mem, prep.ma.nbytes)
-            mag_mem = max(mag_mem, prep.mag.nbytes)
+
+            log(4, 'Free memory on device: %.2f GB' % (float(cuda.mem_get_info()[0])/1e9))
+            self.ex_data.add_data_block()
+            self.ma_data.add_data_block()
+            self.mag_data.add_data_block()
         
-        # now check remaining memory and allocate as many blocks as would fit
-        mem = cuda.mem_get_info()[0]
-        if self.ex_data is not None:
-            mem += self.ex_data.memory    # as we realloc these, consider as free memory
-        if self.ma_data is not None:
-            mem += self.ma_data.memory
-        if self.mag_data is not None:
-            mem += self.mag_data.memory
-
-        blk = ex_mem * EX_MA_BLOCKS_RATIO + ma_mem + mag_mem
-        fit = int(mem - 200*1024*1024) // blk  # leave 200MB room for safety
-        fit = min(MAX_BLOCKS, fit)
-        nex = min(fit * EX_MA_BLOCKS_RATIO, blocks)
-        nma = min(fit, blocks)
-        nstreams = min(MAX_STREAMS, blocks)
-
-        log(4, 'PyCUDA blocks fitting on GPU: exit arrays={}, ma_arrays={}, streams={}, totalblocks={}'.format(nex, nma, nstreams, blocks))
-        # reset memory or create new
-        if self.ex_data is not None:
-            self.ex_data.reset(ex_mem, nex)
-        else:
-            self.ex_data = GpuDataManager(ex_mem, nex, True)
-        if self.ma_data is not None:
-            self.ma_data.reset(ma_mem, nma)
-        else:
-            self.ma_data = GpuDataManager(ma_mem, nma, False)
-        if self.mag_data is not None:
-            self.mag_data.reset(mag_mem, nma)
-        else:
-            self.mag_data = GpuDataManager(mag_mem, nma, False)
-        self.streams = [GpuStreamData(self.ex_data, self.ma_data, self.mag_data) for _ in range(nstreams)]
-
     def engine_iterate(self, num=1):
         """
         Compute one iteration.
         """
+        # ma_buf = ma_c = np.zeros(FUK.fshape, dtype=np.float32)
         self.dID_list = list(self.di.S.keys())
-
-        prev_event = None
-
-        # atomics or tiled version for probe / object update kernels
         atomics_probe = self.p.probe_update_cuda_atomics
         atomics_object = self.p.object_update_cuda_atomics
         use_tiles = (not atomics_object) or (not atomics_probe)
-
+        
         for it in range(num):
 
             error = {}
@@ -258,9 +162,6 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                 do_update_fourier = (inner == 0)
 
                 # initialize probe and object buffer to receive an update
-                # we do this on the first stream we work on
-                streamdata = self.streams[self.cur_stream]
-                streamdata.start_compute(prev_event)
                 if do_update_object:
                     for oID, ob in self.ob.storages.items():
                         cfact = self.ob_cfact[oID]
@@ -268,30 +169,21 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                         obb = self.ob_buf.S[oID]
 
                         if self.p.obj_smooth_std is not None:
-                            log(4,'Smoothing object, cfact is %.2f' % cfact)
+                            log(4, 'Smoothing object, cfact is %.2f' % cfact)
                             smooth_mfs = [self.p.obj_smooth_std, self.p.obj_smooth_std]
                             # We need a third copy, because we still need ob.gpu for the fourier update
-                            # obb.gpu[:] = ob.gpu[:]
-                            cuda.memcpy_dtod_async(dest=obb.gpu.ptr,
-                                                   src=ob.gpu.ptr,
-                                                   size=ob.gpu.nbytes,
-                                                   stream=streamdata.queue)
-                            streamdata.queue.synchronize()
-                            self.GSK.queue = streamdata.queue
+                            obb.gpu[:] = ob.gpu[:]
                             self.GSK.convolution(obb.gpu, smooth_mfs, tmp=obb.tmp)
-                            obb.gpu._axpbz(np.complex64(cfact), 0, obb.gpu, stream=streamdata.queue)
+                            obb.gpu *= np.complex64(cfact)
                         else:
-                            ob.gpu._axpbz(np.complex64(cfact), 0, obb.gpu, stream=streamdata.queue)
-                        obn.gpu.fill(np.float32(cfact), stream=streamdata.queue)
-                
-                self.ex_data.syncback = True
+                            # obb.gpu[:] = ob.gpu * np.complex64(cfact)
+                            ob.gpu._axpbz(np.complex64(cfact), 0, obb.gpu)
+                        obn.gpu.fill(np.float32(cfact))
 
                 # First cycle: Fourier + object update
-                for dID in self.dID_list:
-                    t1 = time.time()
+                for iblock, dID in enumerate(self.dID_list):
                     prep = self.diff_info[dID]
-                    streamdata = self.streams[self.cur_stream]
-                    
+
                     # find probe, object in exit ID in dependence of dID
                     pID, oID, eID = prep.poe_IDs
 
@@ -304,13 +196,6 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                     pbound = self.pbound_scan[prep.label]
                     aux = kern.aux
                     PROP = kern.PROP
-
-                    # set streams
-                    queue = streamdata.queue
-                    FUK.queue = queue
-                    AWK.queue = queue
-                    POK.queue = queue
-                    PROP.queue = queue
 
                     # get addresses and auxilliary array
                     addr = prep.addr_gpu
@@ -326,105 +211,97 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                     obb = self.ob_buf.S[oID].gpu
                     pr = self.pr.S[pID].gpu
 
-                    # transfer exit wave to gpu
-                    ex = streamdata.ex_to_gpu(dID, prep.ex)
-                    
-                    # transfer ma/mag data to gpu if needed
-                    if do_update_fourier:
-                        # transfer other input data in
-                        ma, mag = streamdata.ma_to_gpu(dID, prep.ma, prep.mag)
-                    # waits for compute on previous stream to finish before continuing
-                    streamdata.start_compute(prev_event)
+                    # Schedule ex to device
+                    ev_ex, ex, data_ex = self.ex_data.to_gpu(prep.ex, dID, self.qu_htod)
 
-                    # Fourier update
+                    # Fourier update.
                     if do_update_fourier:
-                        log(4, '------ Fourier update -----', True)
+                        self.ex_data.syncback = True
+                        log(4, '----- Fourier update -----', True)
+
+                        # Schedule ma & mag to device
+                        ev_ma, ma, data_ma = self.ma_data.to_gpu(prep.ma, dID, self.qu_htod)
+                        ev_mag, mag, data_mag = self.mag_data.to_gpu(prep.mag, dID, self.qu_htod)
 
                         ## compute log-likelihood
                         if self.p.compute_log_likelihood:
-                            t1 = time.time()
                             AWK.build_aux_no_ex(aux, addr, ob, pr)
                             PROP.fw(aux, aux)
-                            FUK.log_likelihood(aux, addr, mag, ma, err_phot)                    
-                            self.benchmark.F_LLerror += time.time() - t1
+                            # synchronize h2d stream with compute stream
+                            self.queue.wait_for_event(ev_mag)
+                            FUK.log_likelihood(aux, addr, mag, ma, err_phot)
 
-                        ## prep + forward FFT
-                        t1 = time.time()
+                        # synchronize h2d stream with compute stream
+                        self.queue.wait_for_event(ev_ex)
                         #AWK.build_aux(aux, addr, ob, pr, ex, alpha=self.p.alpha)
                         AWK.make_aux(aux, addr, ob, pr, ex, c_po=self._c, c_e=1-self._c)
-                        self.benchmark.A_Build_aux += time.time() - t1
 
-                        t1 = time.time()
+                        ## FFT
                         PROP.fw(aux, aux)
-                        self.benchmark.B_Prop += time.time() - t1
 
                         ## Deviation from measured data
-                        t1 = time.time()
+                        # synchronize h2d stream with compute stream
+                        self.queue.wait_for_event(ev_mag)
                         FUK.fourier_error(aux, addr, mag, ma, ma_sum)
                         FUK.error_reduce(addr, err_fourier)
                         FUK.fmag_all_update(aux, addr, mag, ma, err_fourier, pbound)
-                        self.benchmark.C_Fourier_update += time.time() - t1
-                        streamdata.record_done_ma(dID)
 
-                        ## Backward FFT
-                        t1 = time.time()
+                        data_mag.record_done(self.queue, 'compute')
+                        data_ma.record_done(self.queue, 'compute')
+
                         PROP.bw(aux, aux)
                         ## apply changes
                         #AWK.build_exit(aux, addr, ob, pr, ex, alpha=self.p.alpha)
                         AWK.make_exit(aux, addr, ob, pr, ex, c_a=self._b, c_po=self._a, c_e=-(self._a + self._b))
                         FUK.exit_error(aux, addr)
                         FUK.error_reduce(addr, err_exit)
-                        self.benchmark.E_Build_exit += time.time() - t1
-                        
-                        self.benchmark.calls_fourier += 1
 
                     prestr = '%d Iteration (Overlap) #%02d:  ' % (parallel.rank, inner)
 
-                    # Object update
+                    # Update object
                     if do_update_object:
                         log(4, prestr + '----- object update -----', True)
-                        t1 = time.time()
-
                         addrt = addr if atomics_object else addr2
+                        self.queue.wait_for_event(ev_ex)
                         POK.ob_update(addrt, obb, obn, pr, ex, atomics=atomics_object)
-                        self.benchmark.object_update += time.time() - t1
-                        self.benchmark.calls_object += 1
 
-                    # end_compute is to allow aux + ob re-use, so we can mark it here
-                    prev_event = streamdata.end_compute()
-                    streamdata.record_done_ex(dID)
-                    self.cur_stream = (self.cur_stream + self.stream_direction) % len(self.streams)
+                    data_ex.record_done(self.queue, 'compute')
+                    if iblock + len(self.ex_data) < len(self.dID_list):
+                        data_ex.from_gpu(self.qu_dtoh)
 
-                # swap direction for next time
-                if do_update_fourier:
+                # swap direction
+                if do_update_fourier or do_update_object:
                     self.dID_list.reverse()
-                    self.stream_direction = -self.stream_direction
-                    # make sure we start with the same stream were we stopped
-                    self.cur_stream = (self.cur_stream + self.stream_direction) % len(self.streams)
+
+                # wait for compute stream to finish
+                self.queue.synchronize()
 
                 if do_update_object:
-                    self._object_allreduce()
 
-                # Exit if probe should fnot yet be updated
+                    for oID, ob in self.ob.storages.items():
+                        obn = self.ob_nrm.S[oID]
+                        obb = self.ob_buf.S[oID]
+                        self.multigpu.allReduceSum(obb.gpu)
+                        self.multigpu.allReduceSum(obn.gpu)
+                        obb.gpu /= obn.gpu
+
+                        self.clip_object(obb.gpu)
+                        ob.gpu[:] = obb.gpu
+
+                # Exit if probe should not yet be updated
                 if not do_update_probe:
                     break
+                self.ex_data.syncback = False
 
                 # Update probe
                 log(4, prestr + '----- probe update -----', True)
-                self.ex_data.syncback = False
                 change = self.probe_update()
-                
-                # swap direction for next time
-                self.dID_list.reverse()
-                self.stream_direction = -self.stream_direction
-                # make sure we start with the same stream were we stopped
-                self.cur_stream = (self.cur_stream + self.stream_direction) % len(self.streams)
-
                 log(4, prestr + 'change in probe is %.3f' % change, True)
 
                 # stop iteration if probe change is small
                 if change < self.p.overlap_converge_factor: break
 
+            self.queue.synchronize()
             parallel.barrier()
 
             if self.do_position_refinement and (self.curiter):
@@ -437,9 +314,7 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                     Iterates through all positions and refines them by a given algorithm. 
                     """
                     log(4, "----------- START POS REF -------------")
-                    prev_event = None
                     for dID in self.di.S.keys():
-                        streamdata = self.streams[self.cur_stream]
 
                         prep = self.diff_info[dID]
                         pID, oID, eID = prep.poe_IDs
@@ -451,16 +326,16 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                         original_addr = prep.original_addr
                         mangled_addr = prep.mangled_addr_gpu
                         ma_sum = prep.ma_sum_gpu
-                        ma, mag = streamdata.ma_to_gpu(dID, prep.ma, prep.mag)
                         err_fourier = prep.err_fourier_gpu
                         error_state = prep.error_state_gpu
 
                         PCK = kern.PCK
                         TK = kern.TK
                         PROP = kern.PROP
-                        PCK.queue = streamdata.queue
-                        TK.queue = streamdata.queue
-                        PROP.queue = streamdata.queue
+
+                        # Make sure our data arrays are on device
+                        ev_ma, ma, data_ma = self.ma_data.to_gpu(prep.ma, dID, self.qu_htod)
+                        ev_mag, mag, data_mag = self.mag_data.to_gpu(prep.mag, dID, self.qu_htod)
 
                         # Keep track of object boundaries
                         max_oby = ob.shape[-2] - aux.shape[-2] - 1
@@ -469,6 +344,10 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                         # We need to re-calculate the current error
                         PCK.build_aux(aux, addr, ob, pr)
                         PROP.fw(aux, aux)
+                        # wait for data to arrive
+                        self.queue.wait_for_event(ev_mag)
+
+                        # We need to re-calculate the current error 
                         if self.p.position_refinement.metric == "fourier":
                             PCK.fourier_error(aux, addr, mag, ma, ma_sum)
                             PCK.error_reduce(addr, err_fourier)
@@ -476,14 +355,11 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                             PCK.log_likelihood(aux, addr, mag, ma, err_fourier)
                         cuda.memcpy_dtod_async(dest=error_state.ptr,
                                                src=err_fourier.ptr,
-                                               size=err_fourier.nbytes,
-                                               stream=streamdata.queue)
-                        streamdata.start_compute(prev_event)
+                                               size=err_fourier.nbytes, stream=self.queue)
 
                         log(4, 'Position refinement trial: iteration %s' % (self.curiter))
                         PCK.mangler.setup_shifts(self.curiter, nframes=addr.shape[0])
                         for i in range(PCK.mangler.nshifts):
-                            streamdata.queue.synchronize()
                             PCK.mangler.get_address(i, addr, mangled_addr, max_oby, max_obx)
                             PCK.build_aux(aux, mangled_addr, ob, pr)
                             PROP.fw(aux, aux)
@@ -494,32 +370,27 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
                                 PCK.log_likelihood(aux, mangled_addr, mag, ma, err_fourier)
                             PCK.update_addr_and_error_state(addr, error_state, mangled_addr, err_fourier)
 
+                        data_mag.record_done(self.queue, 'compute')
+                        data_ma.record_done(self.queue, 'compute')
                         cuda.memcpy_dtod_async(dest=err_fourier.ptr,
                                                src=error_state.ptr,
-                                               size=err_fourier.nbytes,
-                                               stream=streamdata.queue)
+                                               size=err_fourier.nbytes, stream=self.queue)
                         if use_tiles:
                             s1 = prep.addr_gpu.shape[0] * prep.addr_gpu.shape[1]
                             s2 = prep.addr_gpu.shape[2] * prep.addr_gpu.shape[3]
                             TK.transpose(prep.addr_gpu.reshape(s1, s2), prep.addr2_gpu.reshape(s2, s1))
 
-                        prev_event = streamdata.end_compute()
-                        
-                        # next stream
-                        self.cur_stream = (self.cur_stream + self.stream_direction) % len(self.streams)
-
             self.curiter += 1
+            self.queue.synchronize()
 
-        #print('end loop, syncall and copy back')
-        for sd in self.streams:
-            sd.synchronize()
-        
         for name, s in self.ob.S.items():
-            s.gpu.get(s.data)
+            s.data[:] = s.gpu.get()
         for name, s in self.pr.S.items():
-            s.gpu.get(s.data)
+            s.data[:] = s.gpu.get()
 
-        # FIXXME: copy to pinned memory
+        # costly but needed to sync back with
+        # for name, s in self.ex.S.items():
+        #     s.data[:] = s.gpu.get()
         for dID, prep in self.diff_info.items():
             err_fourier = prep.err_fourier_gpu.get()
             err_phot = prep.err_phot_gpu.get()
@@ -529,69 +400,50 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
 
         self.error = error
         return error
-    
-    def _object_allreduce(self):
-        # make sure that all transfers etc are finished
-        for sd in self.streams:
-            sd.synchronize()
-        # sync all
-        for oID, ob in self.ob.storages.items():
-            obn = self.ob_nrm.S[oID]
-            obb = self.ob_buf.S[oID]
-            self.multigpu.allReduceSum(obb.gpu)
-            self.multigpu.allReduceSum(obn.gpu)
-            obb.gpu /= obn.gpu
-            
-            self.clip_object(obb.gpu)
-            ob.gpu[:] = obb.gpu
 
     ## probe update
     def probe_update(self, MPI=False):
-        t1 = time.time()
-        streamdata = self.streams[self.cur_stream]
+        queue = self.queue
         use_atomics = self.p.probe_update_cuda_atomics
         # storage for-loop
         change_gpu = gpuarray.zeros((1,), dtype=np.float32)
-        prev_event = None
         for pID, pr in self.pr.storages.items():
             prn = self.pr_nrm.S[pID]
             cfact = self.pr_cfact[pID]
-            pr.gpu._axpbz(np.complex64(cfact), 0, pr.gpu, stream=streamdata.queue)
-            prn.gpu.fill(np.float32(cfact), stream=streamdata.queue)
+            # pr.gpu *= np.float64(cfact)
+            pr.gpu._axpbz(np.complex64(cfact), 0, pr.gpu, stream=queue)
+            prn.gpu.fill(np.float32(cfact), stream=self.queue)
 
-        for dID in self.dID_list:
+        for iblock, dID in enumerate(self.dID_list):
             prep = self.diff_info[dID]
-            streamdata = self.streams[self.cur_stream]
 
             POK = self.kernels[prep.label].POK
-            POK.queue = streamdata.queue
             # find probe, object in exit ID in dependence of dID
             pID, oID, eID = prep.poe_IDs
 
-            ex = streamdata.ex_to_gpu(dID, prep.ex)
+            ev, ex, data_ex = self.ex_data.to_gpu(prep.ex, dID, self.qu_htod)
+            self.queue.wait_for_event(ev)
 
-            # scan for-loop
             addrt = prep.addr_gpu if use_atomics else prep.addr2_gpu
-            streamdata.start_compute(prev_event)
             ev = POK.pr_update(addrt,
                                self.pr.S[pID].gpu,
                                self.pr_nrm.S[pID].gpu,
                                self.ob.S[oID].gpu,
                                ex,
                                atomics=use_atomics)
-            streamdata.record_done_ex(dID)
-            prev_event = streamdata.end_compute()
-            self.cur_stream = (self.cur_stream + self.stream_direction) % len(self.streams)
 
-        # sync all streams first
-        for sd in self.streams:
-            sd.synchronize()
+            data_ex.record_done(self.queue, 'compute')
+            if iblock + len(self.ex_data) < len(self.dID_list):
+                data_ex.from_gpu(self.qu_dtoh)
 
+        self.dID_list.reverse()
+
+        self.queue.synchronize()
         for pID, pr in self.pr.storages.items():
 
             buf = self.pr_buf.S[pID]
             prn = self.pr_nrm.S[pID]
-            
+
             self.multigpu.allReduceSum(pr.gpu)
             self.multigpu.allReduceSum(prn.gpu)
             pr.gpu /= prn.gpu
@@ -605,26 +457,25 @@ class _ProjectionEngine_pycuda_streams(projectional_pycuda._ProjectionEngine_pyc
             self.multigpu.allReduceSum(change_gpu)
             change = change_gpu.get().item() / parallel.size
 
-        # print 'probe update: ' + str(time.time()-t1)
-        self.benchmark.probe_update += time.time() - t1
-        self.benchmark.calls_probe += 1
-
         return np.sqrt(change)
 
     def engine_finalize(self, benchmark=False):
         """
         Clear all GPU data, pinned memory, etc
-        """ 
-        self.streams = None
+        """
         self.ex_data = None
         self.ma_data = None
         self.mag_data = None
+
+        # copy data to cpu
+        for name, s in self.pr.S.items():
+            s.data = np.copy(s.data)  # is this the same as s.data.get()?
 
         super().engine_finalize(benchmark)
 
 
 @register()
-class DM_pycuda_streams(_ProjectionEngine_pycuda_streams, DMMixin):
+class DM_pycuda(_ProjectionEngine_pycuda, DMMixin):
     """
     A full-fledged Difference Map engine accelerated with pycuda.
 
@@ -639,13 +490,13 @@ class DM_pycuda_streams(_ProjectionEngine_pycuda_streams, DMMixin):
     """
 
     def __init__(self, ptycho_parent, pars=None):
-        _ProjectionEngine_pycuda_streams.__init__(self, ptycho_parent, pars)
+        _ProjectionEngine_pycuda.__init__(self, ptycho_parent, pars)
         DMMixin.__init__(self, self.p.alpha)
         ptycho_parent.citations.add_article(**self.article)
 
 
 @register()
-class RAAR_pycuda_streams(_ProjectionEngine_pycuda_streams, RAARMixin):
+class RAAR_pycuda(_ProjectionEngine_pycuda, RAARMixin):
     """
     A RAAR engine in accelerated with pycuda.
 
@@ -661,5 +512,5 @@ class RAAR_pycuda_streams(_ProjectionEngine_pycuda_streams, RAARMixin):
 
     def __init__(self, ptycho_parent, pars=None):
 
-        _ProjectionEngine_pycuda_streams.__init__(self, ptycho_parent, pars)
+        _ProjectionEngine_pycuda.__init__(self, ptycho_parent, pars)
         RAARMixin.__init__(self, self.p.beta)
